Exploded the text module into multiple files
