+++ golem-cli/src/model/text/account.rs
@@ -1,0 +1,89 @@
+// Copyright 2024-2025 Golem Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::model::text::fmt::*;
+use golem_cloud_client::model::{Account, Role};
+use serde::{Deserialize, Serialize};
+
+fn account_fields(account: &Account) -> Vec<(String, String)> {
+    let mut fields = FieldsBuilder::new();
+
+    fields
+        .fmt_field("Account ID", &account.id, format_main_id)
+        .fmt_field("E-mail", &account.email, format_id)
+        .field("Name", &account.name);
+
+    fields.build()
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AccountGetView(pub Account);
+
+impl MessageWithFields for AccountGetView {
+    fn message(&self) -> String {
+        format!(
+            "Got metadata for account {}",
+            format_message_highlight(&self.0.id)
+        )
+    }
+
+    fn fields(&self) -> Vec<(String, String)> {
+        account_fields(&self.0)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AccountNewView(pub Account);
+
+impl MessageWithFields for AccountNewView {
+    fn message(&self) -> String {
+        format!(
+            "Created new account {}",
+            format_message_highlight(&self.0.id)
+        )
+    }
+
+    fn fields(&self) -> Vec<(String, String)> {
+        account_fields(&self.0)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AccountUpdateView(pub Account);
+
+impl MessageWithFields for AccountUpdateView {
+    fn message(&self) -> String {
+        format!("Updated account {}", format_message_highlight(&self.0.id))
+    }
+
+    fn fields(&self) -> Vec<(String, String)> {
+        account_fields(&self.0)
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
+pub struct GrantGetView(pub Vec<Role>);
+
+impl TextView for GrantGetView {
+    fn log(&self) {
+        if self.0.is_empty() {
+            println!("No roles granted")
+        } else {
+            println!("Granted roles:");
+            for role in &self.0 {
+                println!("  - {}", role);
+            }
+        }
+    }
+}
