+++ golem-cli/src/model/text/mod.rs
@@ -1,0 +1,29 @@
+// Copyright 2024-2025 Golem Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+pub mod account;
+pub mod api_definition;
+pub mod api_deployment;
+pub mod api_domain;
+pub mod api_security;
+pub mod certificate;
+pub mod component;
+pub mod fmt;
+pub mod help;
+pub mod plugin;
+pub mod profile;
+pub mod project;
+pub mod template;
+pub mod token;
+pub mod worker;
